Refactor backend chat route
